way-back-home/level_5/pyproject.toml
[project]
name = "waybackhome-level5"
version = "0.1.0"
description = "Way Back Home - Level 5"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-genai==1.56.0",
    "websockets==15.0.1",
    "fastapi==0.128.3",
    "python-dotenv==1.2.1",
    "google-adk==1.24.1",
    "uvicorn==0.40.0",
    "numpy>=2.2.0",
    "aiokafka==0.13.0",
    "a2a-sdk[kafka]>=0.3.24",
    "sse-starlette>=2.0.0",
    "pydantic>=2.11.7",
]
